Move code from multiple headers to .c files
src/arch/xtensa/include/arch/task.h, src/include/sof/wait.h,
and src/include/sof/panic.h contained functions, too big for
inlining. Move them out to .c files.
