+++ src/lib/panic.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016, Intel Corporation
+ * Copyright (c) 2018, Intel Corporation
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -28,42 +28,70 @@
  * Author: Liam Girdwood <liam.r.girdwood@linux.intel.com>
  */
 
-#include <platform/memory.h>
+#include <sof/alloc.h>
 #include <sof/interrupt.h>
-#include <platform/interrupt.h>
 #include <sof/mailbox.h>
-#include <arch/init.h>
-#include <arch/task.h>
-#include <sof/init.h>
-#include <sof/lock.h>
 #include <sof/panic.h>
+#include <sof/sof.h>
+#include <sof/trace.h>
+
+#include <platform/platform.h>
+
+#include <uapi/ipc/trace.h>
+
 #include <stdint.h>
+#include <stdlib.h>
 
-/**
- * \file arch/xtensa/up/init.c
- * \brief Xtensa UP initialization functions
- * \authors Liam Girdwood <liam.r.girdwood@linux.intel.com>
- */
-
-#if DEBUG_LOCKS
-/** \brief Debug lock. */
-uint32_t lock_dbg_atomic = 0;
-
-/** \brief Debug locks per user. */
-uint32_t lock_dbg_user[DBG_LOCK_USERS] = {0};
-#endif
-
-/**
- * \brief Initializes architecture.
- * \param[in,out] sof Firmware main context.
- * \return Error status.
- */
-int arch_init(struct sof *sof)
+void dump_panicinfo(void *addr, struct sof_ipc_panic_info *panic_info)
 {
-	register_exceptions();
-	arch_assign_tasks();
-	return 0;
+	if (!panic_info)
+		return;
+	rmemcpy(addr, panic_info, sizeof(struct sof_ipc_panic_info));
+	dcache_writeback_region(addr, sizeof(struct sof_ipc_panic_info));
 }
 
-int slave_core_init(struct sof *sof) { return 0; }
+/* panic and rewind stack */
+void panic_rewind(uint32_t p, uint32_t stack_rewind_frames,
+		  struct sof_ipc_panic_info *panic_info)
+{
+	void *ext_offset;
+	size_t count;
+	uint32_t oldps;
 
+	/* disable all IRQs */
+	oldps = interrupt_global_disable();
+
+	/* dump DSP core registers */
+	ext_offset = arch_dump_regs(oldps);
+
+	/* dump panic info, filename ane linenum */
+	dump_panicinfo(ext_offset, panic_info);
+	ext_offset += sizeof(struct sof_ipc_panic_info);
+
+	count = MAILBOX_EXCEPTION_SIZE -
+		(size_t)(ext_offset - mailbox_get_exception_base());
+	/* dump stack frames */
+	p = dump_stack(p, ext_offset, stack_rewind_frames, count);
+
+	/* panic - send IPC oops message to host */
+	platform_panic(p);
+
+	/* flush last trace messages */
+	trace_flush();
+
+	/* and loop forever */
+	while (1)
+		;
+}
+
+void __panic(uint32_t p, char *filename, uint32_t linenum)
+{
+	struct sof_ipc_panic_info panicinfo;
+	int strlen;
+
+	strlen = rstrlen(filename);
+	panicinfo.linenum = linenum;
+	rmemcpy(panicinfo.filename, filename, strlen + 1);
+
+	panic_rewind(p, 0, &panicinfo);
+}
