+++ src/platform/apollolake/include/platform/clk.h
@@ -34,6 +34,7 @@
 
 #include <sof/cpu.h>
 #include <sof/io.h>
+#include <platform/shim.h>
 
 #define CLK_CPU(x)	(x)
 #define CLK_SSP		2
