+++ src/platform/baytrail/include/platform/platform.h
@@ -34,6 +34,7 @@
 #define __PLATFORM_PLATFORM_H__
 
 #include <sof/platform.h>
+#include <platform/clk.h>
 #include <platform/shim.h>
 #include <platform/interrupt.h>
