+++ src/platform/haswell/include/platform/clk.h
@@ -32,6 +32,7 @@
 #define __PLATFORM_CLOCK__
 
 #include <sof/io.h>
+#include <platform/shim.h>
 
 #define CLK_CPU(x)	(x)
 #define CLK_SSP		1
