+++ src/platform/haswell/include/platform/platform.h
@@ -33,6 +33,7 @@
 #define __PLATFORM_PLATFORM_H__
 
 #include <sof/platform.h>
+#include <platform/clk.h>
 #include <platform/shim.h>
 #include <platform/interrupt.h>
