+++ src/platform/icelake/include/platform/clk.h
@@ -35,6 +35,7 @@
 
 #include <sof/cpu.h>
 #include <sof/io.h>
+#include <platform/shim.h>
 
 #define CLK_CPU(x)	(x)
 #define CLK_SSP		4
