+++ src/platform/intel/cavs/platform.c
@@ -52,6 +52,7 @@
 #include <sof/io.h>
 #include <sof/trace.h>
 #include <sof/audio/component.h>
+#include <sof/drivers/timer.h>
 #include <sof/cpu.h>
 #include <sof/notifier.h>
 #include <sof/spi.h>
